mock data populated

+++ controllers/student-controller.js
@@ -1,12 +1,15 @@
 const Student = require('../models/student')
 const Logger = require('../models/logger')
 const { StatusCodes } = require('http-status-codes')
-const { NotFoundError } = require('../errors')
+const { NotFoundError, BadRequestError } = require('../errors')
 const getAllStudents = async (req, res) => {
   const students = await Student.find({})
   res.status(StatusCodes.OK).json({ students })
 }
 const createStudent = async (req, res) => {
+  const { name, dob, mobile, email } = req.body
+  if (!name || !dob || !mobile || !email)
+    throw new BadRequestError('Please provide valid credentials')
   const student = await Student.create(req.body)
   const log = await Logger.create({
     message: `${student.name} is added in database`,
@@ -15,7 +18,9 @@ const createStudent = async (req, res) => {
 }
 const updateStudent = async (req, res) => {
   let { id: updateId } = req.params
-
+  const { name, dob, mobile, email } = req.body
+  if (!updateId || !name || !dob || !mobile || !email)
+    throw new BadRequestError('Please provide valid credentials')
   const student = await Student.findOneAndUpdate(
     {
       _id: updateId,
@@ -23,7 +28,7 @@ const updateStudent = async (req, res) => {
     req.body,
     { new: true, runValidators: true }
   )
-  if (!student) throw new NotFoundError('no student')
+  if (!student) throw new NotFoundError('No student in database with this id')
   const log = await Logger.create({
     message: `${student.name} is updated in database`,
   })
@@ -31,8 +36,9 @@ const updateStudent = async (req, res) => {
 }
 const deleteStudent = async (req, res) => {
   let { id: studentId } = req.params
+  if (!studentId) throw new BadRequestError('Invalid credentials')
   const student = await Student.findOneAndDelete({ _id: studentId })
-  if (!student) throw new NotFoundError('no student')
+  if (!student) throw new NotFoundError('No student in database with this id')
   const log = await Logger.create({
     message: `${student.name} is deleted from database`,
   })
